separate positional args from Config definition
Until now the user had to define their own `args` field in their Config
struct, leading to a lot more comptime validation being needed. Since
we know we need to return positional arguments in that format, we don't
need to have the user specify it, and simply return it alongside their
Config struct.